requirements.txt
nltk==3.9.1
wikipedia==1.4.0
python-docx==1.1.2
reportlab==4.1.0
tqdm==4.66.5
names==0.3.0
beautifulsoup4==4.12.2
pytest==7.4.0
flake8==6.1.0